% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_abundances_methods.R
\docType{methods}
\name{abundances}
\alias{abundances}
\alias{abundances,otu_table-method}
\alias{abundances,}
\alias{otu_table-method}
\alias{abundances,phyloseq-method}
\alias{abundances,microbiomeMarker-method}
\alias{abundances,assay-method}
\alias{assay-method}
\alias{abundances,SummarizedExperiment-method}
\title{Extract taxa abundances}
\usage{
abundances(
  object,
  transform = c("identity", "log10", "log10p", "log2", "log2p", "SquareRoot",
    "CubicRoot", "logit"),
  norm = FALSE
)

\S4method{abundances}{otu_table}(
  object,
  transform = c("identity", "log10", "log10p", "log2", "log2p", "SquareRoot",
    "CubicRoot", "logit"),
  norm = FALSE
)

\S4method{abundances}{phyloseq}(
  object,
  transform = c("identity", "log10", "log10p", "log2", "log2p", "SquareRoot",
    "CubicRoot", "logit"),
  norm = FALSE
)

\S4method{abundances}{microbiomeMarker}(
  object,
  transform = c("identity", "log10", "log10p", "log2", "log2p", "SquareRoot",
    "CubicRoot", "logit")
)

\S4method{abundances}{assay}(
  object,
  transform = c("identity", "log10", "log10p", "log2", "log2p", "SquareRoot",
    "CubicRoot", "logit"),
  norm = FALSE
)

\S4method{abundances}{SummarizedExperiment}(
  object,
  transform = c("identity", "log10", "log10p", "log2", "log2p", "SquareRoot",
    "CubicRoot", "logit"),
  norm = FALSE
)
}
\arguments{
\item{object}{\code{\linkS4class{otu_table}}, \code{\linkS4class{phyloseq}}, or
\code{\linkS4class{microbiomeMarker}}.}

\item{transform}{transformation to apply, the options inclulde:
\itemize{
\item "identity", return the original data without any transformation.
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
\item "log2", the transformation is \code{log2(object)}, and if the data contains
zeros the transformation is \code{log2(1 + object)}.
\item "log2p", the transformation is \code{log2(1 + object)}.
\item "SquareRoot", the transformation is \verb{Square Root}.
\item "CubicRoot", the transformation is \verb{Cubic Root}.
\item "logit", the transformation is \verb{Zero-inflated Logit Transformation}
(Does not work well for microbiome data).
}}

\item{norm}{logical, indicating whether or not to return the normalized
taxa abundances.}
}
\value{
abundance matrix with taxa in rows and samples in columns.
}
\description{
Extract taxa abundances from phyloseq objects.
}
\examples{
data(caporaso)
abd <- abundances(caporaso)
}
\seealso{
\code{\linkS4class{otu_table}}, \code{\linkS4class{phyloseq}},
\code{\linkS4class{microbiomeMarker}},\code{\link{transform_abundances}}
}
