% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{calculate_mean_abundance}
\alias{calculate_mean_abundance}
\title{Calculate the Mean abundance of features per group (two groups)}
\usage{
calculate_mean_abundance(profile, metadata, groups)
}
\arguments{
\item{profile}{(Required). matrix.}

\item{metadata}{(Required). matrix.}

\item{groups}{(Required). character.}
}
\description{
Calculate the Mean abundance of features per group (two groups)
}
