% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{calculate_MeanRank_abundance}
\alias{calculate_MeanRank_abundance}
\title{Calculate the Rank Mean abundance of features per group(two groups)}
\usage{
calculate_MeanRank_abundance(profile, metadata, groups)
}
\arguments{
\item{profile}{(Required). matrix.}

\item{metadata}{(Required). matrix.}

\item{groups}{(Required). character.}
}
\description{
Calculate the Rank Mean abundance of features per group(two groups)
}
