% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{calculate_median_abundance}
\alias{calculate_median_abundance}
\title{Calculate the Median abundance of features per group (two groups)}
\usage{
calculate_median_abundance(profile, metadata, groups)
}
\arguments{
\item{profile}{matrix; (Required).}

\item{metadata}{matrix; (Required).}

\item{groups}{character; (Required).}
}
\description{
Calculate the Median abundance of features per group (two groups)
}
