% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{calculate_occurrence_taxa}
\alias{calculate_occurrence_taxa}
\title{Calculate the Occurrence of features per group (two groups)}
\usage{
calculate_occurrence_taxa(profile, metadata, groups)
}
\arguments{
\item{profile}{(Required). matrix.}

\item{metadata}{(Required). matrix.}

\item{groups}{(Required). character.}
}
\description{
Calculate the Occurrence of features per group (two groups)
}
