% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_phyloseq}
\alias{check_phyloseq}
\title{check whether the object is phyloseq-class object}
\usage{
check_phyloseq(ps, fill_na_taxa = TRUE)
}
\arguments{
\item{ps}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object.}

\item{fill_na_taxa}{(Optional). whether there is taxa table (default: TRUE).}
}
\description{
check whether the object is phyloseq-class object
}
