% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_rank_names}
\alias{check_rank_names}
\title{check whether all names of taxonomic ranks include in available_ranks}
\usage{
check_rank_names(ps)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object; (Required).}
}
\description{
check whether all names of taxonomic ranks include in available_ranks
}
