% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_sample_names}
\alias{check_sample_names}
\title{check whether samples' names start with numeric and then paste "S_"}
\usage{
check_sample_names(object)
}
\arguments{
\item{object}{(Optional). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} or \code{\link[phyloseq:otu_table-class]{phyloseq::otu_table}} or
\code{\link[phyloseq:otu_table-class]{phyloseq::otu_table}} or matrix/metadata.}
}
\description{
check whether samples' names start with numeric and then paste "S_"
}
