% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alldata.R
\docType{data}
\name{data-Zeybel_2022_metabolite}
\alias{data-Zeybel_2022_metabolite}
\alias{Zeybel_2022_metabolite}
\title{fecal metabolites from Zeybel et al. - 2022 paper}
\format{
a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object
}
\description{
The data from Zeybel et al. - 2022 paper, which investigated the
microbiome composition and development of 56 MAFLD,
identifying gut microbiome associations with oral microbiome.
}
\references{
Zeybel, Mujdat, et al. "Multiomics analysis reveals the impact of microbiota
on host metabolism in hepatic steatosis."
Advanced Science 9.11 (2022): 2104373.
}
