% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{default_color}
\alias{default_color}
\title{generate multiple colors for visualization}
\usage{
default_color(number = 74)
}
\arguments{
\item{number}{(Required). integer. the number of colors (default: 74).}
}
\description{
This funciton is to generate multiple colors for visualization
}
