% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_distance.R
\name{.get_GUniFrac}
\alias{.get_GUniFrac}
\title{Generalized UniFrac distances for comparing microbial communities}
\usage{
.get_GUniFrac(otu.tab,
                 tree,
                 alpha)
}
\arguments{
\item{otu.tab}{(Required). a matrix, an OTU count table, row - n sample, column - q OTU.}

\item{tree}{(Required). tree. a rooted phylogenetic tree of R class “phylo”.}

\item{alpha}{(optional). numeric. the parameter controlling weight on abundant lineages.}
}
\value{
a three dimensional array containing all the UniFrac distance matrices.

unifracs: three dimensional array containing the generalized
UniFrac distances, unweighted UniFrac distance and
variance adjusted UniFrac distances.
}
\description{
A generalized version of commonly used UniFrac distances.
The unweighted and weighted UniFrac, and variance-adjusted weighted UniFrac distances are also implemented.
}
\examples{

\dontrun{
data("cid_ying")
otu.tab <- phyloseq::otu_table(cid_ying)
tree <- phyloseq::phy_tree(cid_ying)

unifracs <- .get_GUniFrac(otu.tab, tree, alpha=c(0, 0.5, 1))$unifracs

dw <- unifracs[, , "d_1"]    # Weighted UniFrac
du <- unifracs[, , "d_UW"]   # Unweighted UniFrac
dv <- unifracs[, , "d_VAW"]  # Variance adjusted weighted UniFrac
d0 <- unifracs[, , "d_0"]    # GUniFrac with alpha 0
d5 <- unifracs[, , "d_0.5"]  # GUniFrac with alpha 0.5

}

}
\references{
https://github.com/cran/GUniFrac/blob/master/R/GUniFrac.R
}
\author{
Created by Jun Chen; modified by Hua Zou (5/14/2022 Shenzhen China)
}
