% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_diversity.R
\name{get_alphaindex}
\alias{get_alphaindex}
\title{Global view on Alpha diversity on microbiota data}
\usage{
get_alphaindex(
    ps,
    level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
    indexes = c("all",
      "Observed", "Chao1", "ACE", "Shannon",
      "Simpson", "InvSimpson", "Fisher", "Evenness")
   )
}
\arguments{
\item{ps}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object.}

\item{level}{(Optional). character. taxonomic level to summarize,
default the top level rank of the \code{ps}. taxonomic level(Kingdom, Phylum,
Class, Order, Family, Genus, Species, Strains; default: NULL).}

\item{indexes}{(Optional). character, meaning that all available
alpha-diversity indexes will be included. Alternatively, you can specify
one or more indexes as a character vector of measure names (default: all).
Values must be among those supported:
c("Observed", "Chao1", "ACE", "Shannon", "Simpson",
"InvSimpson", "Fisher", "Evenness").
"Chao1", "ACE" and "Fisher" only supported by
counts matrix (integers).}
}
\value{
A data.frame of alpha index with metadata.
}
\description{
The function is to use \code{phyloseq::estimate_richness()} to calculate
the alpha diversity.
Performs a number of standard alpha diversity estimates, and returns
results as data.frame.
\strong{NOTE}: You must use untrimmed datasets for meaningful results,
as these estimates (and even the "observed" richness) are highly dependent
on the number of singletons. You can always trim the data later on if needed,
just not before using this function.
}
\examples{

\dontrun{

# relative abundance
data("enterotypes_arumugam")
get_alphaindex(
  ps = enterotypes_arumugam,
  level = "Genus",
  indexes = c("Shannon", "Observed"))

# absolute abundance
data("caporaso")
get_alphaindex(
  ps = caporaso,
  level = "Genus",
  indexes = c("Shannon", "Chao1"))
}


}
\references{
https://scienceparkstudygroup.github.io/microbiome-lesson/04-alpha-diversity/index.html
}
\author{
Created by Hua Zou (12/02/2022 Shenzhen China)
}
