% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_eigValue}
\alias{get_eigValue}
\title{Extract and visualize the eigenvalues/variances of dimensions}
\usage{
get_eigValue(X)
}
\arguments{
\item{X}{(Required). Ordination result.}
}
\description{
This funciton is to obtain eigenvalues/variances of dimensions
}
