% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_import_qiime2.R
\name{import_qiime2}
\alias{import_qiime2}
\title{Convert the qiime2 output of dada2 into phyloseq object}
\usage{
import_qiime2(
        otu_qza,
        taxa_qza,
        sam_tab,
        refseq_qza,
        tree_qza)
}
\arguments{
\item{otu_qza}{(Required). character, file path of the feature table from qiime2.}

\item{taxa_qza}{(Optional). character, file path of the taxonomic table from qiime2,
default \code{NULL}.}

\item{sam_tab}{(Optional). character, file path of the sample metadata in tsv format,
default \code{NULL}.}

\item{refseq_qza}{(Optional). character, file path of the representative sequences from
qiime2, default \code{NULL}.}

\item{tree_qza}{(Optional). character, file path of the phylogenetic tree from
qiime2, default \code{NULL}.}
}
\value{
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object.
}
\description{
Import the qiime2 artifacts, including feature table, taxonomic table,
phylogenetic tree, representative sequence and sample metadata into
phyloseq object.
}
\examples{
\dontrun{
otuqza_file <- system.file(
    "extdata", "table.qza",
    package = "MicrobiomeAnalysis"
)
taxaqza_file <- system.file(
    "extdata", "taxonomy.qza",
    package = "MicrobiomeAnalysis"
)
sample_file <- system.file(
    "extdata", "sample-metadata.tsv",
    package = "MicrobiomeAnalysis"
)
treeqza_file <- system.file(
    "extdata", "tree.qza",
    package = "MicrobiomeAnalysis"
)
ps <- import_qiime2(
    otu_qza = otuqza_file, taxa_qza = taxaqza_file,
    sam_tab = sample_file, tree_qza = treeqza_file
)
ps
}
}
\author{
Created by Yang Cao; modified by Hua Zou (5/14/2022 Shenzhen China)
}
