% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{insertCol}
\alias{insertCol}
\title{Insert Row into a Matrix}
\usage{
insertCol(m, c, v = NA, cName = "")
}
\arguments{
\item{m}{(Required). a matrix.}

\item{c}{(Required). numeric. column number where the new column should be inserted.}

\item{v}{(optional). numeric. values for the new column.}

\item{cName}{(optional). character. the name of the new column.}
}
\value{
a matrix with one more column than the provided matrix m.
}
\description{
Insert Row into a Matrix
}
\examples{

\dontrun{
m <- matrix(1:4, 2)
insertCol(m, 2, 5:6)
}

}
\references{
https://github.com/cran/miscTools/
}
\author{
Created by Hua Zou (5/19/2022 Shenzhen China)
}
