% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{insertRow}
\alias{insertRow}
\title{Insert Row into a Matrix}
\usage{
insertRow(m, r, v = NA, rName = "")
}
\arguments{
\item{m}{(Required). a matrix.}

\item{r}{(Required). numeric. row number where the new row should be inserted.}

\item{v}{(optional). numeric. values for the new row.}

\item{rName}{(optional). character. the name of the new row.}
}
\value{
a matrix with one more row than the provided matrix m.
}
\description{
Insert Row into a Matrix
}
\examples{
\dontrun{
m <- matrix(1:4, 2)
insertRow(m, 2, 5:6)
}

}
\references{
https://github.com/cran/miscTools/
}
\author{
Created by Hua Zou (5/19/2022 Shenzhen China)
}
