% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_ANOSIM.R
\name{run_ANOSIM}
\alias{run_ANOSIM}
\title{Analysis of Similarity (ANOSIM)}
\usage{
run_ANOSIM(
   object,
   level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
   variable,
   method = c("unifrac", "wunifrac", "GUniFrac", "bray", "dpcoa", "jsd"),
   seedNum = 123,
   alpha = 0.5)
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object.}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}

\item{variable}{(Required). character. grouping variable for test.}

\item{method}{(Optional). character. Provide one of the currently supported
options. See \code{distanceMethodList} for a detailed list of the supported options
and links to accompanying documentation. Options include:
\itemize{
\item "bray": bray crutis distance.
\item "unifrac" : unweighted UniFrac distance.
\item "wunifrac": weighted-UniFrac distance.
\item "GUniFrac": The variance-adjusted weighted UniFrac distances (default: alpha=0.5).
\item "dpcoa": sample-wise distance used in Double Principle Coordinate Analysis.
\item "jsd": Jensen-Shannon Divergence.
Alternatively, you can provide a character string that defines a custom
distance method, if it has the form described in \code{designdist} (default: "bray").
}}

\item{seedNum}{(Optional). numeric. specify seeds for reproduction (default: 123).}

\item{alpha}{(Optional). numeric. the parameter for "GUniFrac" controlling
weight on abundant lineages (default: 0.5).}
}
\value{
An ANOSIM model.
}
\description{
The \code{run_ANOSIM} function is used to identify the association between
the community and environmental variables by rank.
}
\details{
The \code{run_ANOSIM} function is used to identify the association between
the community and environmental variables by rank, applying the distance
in profile and calculating the R statistic between community and rank variable
by permutation test to determine the significance. It can be applied to both
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} and \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object.
}
\examples{

\dontrun{

# phyloseq object
data("Zeybel_2022_gut")
run_ANOSIM(Zeybel_2022_gut,
   level = "Phylum",
   variable = "LiverFatClass",
   method = "bray")

# SummarizedExperiment object
data("Zeybel_2022_protein")
run_ANOSIM(Zeybel_2022_protein,
   variable = "LiverFatClass",
   method = "bray")
}

}
\references{
Clarke, K. Robert. "Non‐parametric multivariate analyses of
changes in community structure." Australian journal of ecology 18.1
(1993): 117-143.
}
\author{
Created by Hua Zou (5/15/2022 Shenzhen China)
}
