% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_betadisper.R
\name{run_betadisper}
\alias{run_betadisper}
\title{Multivariate homogeneity of groups dispersions (variances)}
\usage{
run_betadisper(
   object,
   level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
   variable,
   type = c("median","centroid"),
   method = c("unifrac", "wunifrac", "GUniFrac", "bray", "dpcoa", "jsd"),
   seedNum = 123,
   alpha = 0.5)
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object.}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}

\item{variable}{(Required). character. grouping variable for test.}

\item{type}{(Optional). character. the type of analysis to perform. Use the
spatial "median" or the group "centroid" (default: "median").}

\item{method}{(Optional). character. Provide one of the currently supported
options. See \code{distanceMethodList} for a detailed list of the supported options
and links to accompanying documentation. Options include:
\itemize{
\item "bray": bray crutis distance.
\item "unifrac" : unweighted UniFrac distance.
\item "wunifrac": weighted-UniFrac distance.
\item "GUniFrac": The variance-adjusted weighted UniFrac distances (default: alpha=0.5).
\item "dpcoa": sample-wise distance used in Double Principle Coordinate Analysis.
\item "jsd": Jensen-Shannon Divergence.
Alternatively, you can provide a character string that defines a custom
distance method, if it has the form described in \code{designdist} (default: "bray").
}}

\item{seedNum}{(Optional). numeric. specify seeds for reproduction (default: 123).}

\item{alpha}{(Optional). numeric. the parameter for "GUniFrac" controlling
weight on abundant lineages (default: 0.5).}
}
\value{
A betadisper model.
}
\description{
The \code{run_betadisper} function is a multivariate analogue of Levene's test
for homogeneity of variances.
}
\details{
One measure of multivariate dispersion (variance) for a group of samples is
to calculate the average distance of group members to the group centroid or
spatial median (both referred to as 'centroid' from now on unless stated
otherwise) in multivariate space. To test if the dispersions (variances)
of one or more groups are different, the distances of group members to
the group centroid are subject to ANOVA. This is a multivariate analogue
of Levene's test for homogeneity of variances if the distances between
group members and group centroids is the Euclidean distance.
See \code{?vegan::betadisper} for more details. It can be applied to both
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} and \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object.
}
\examples{

\dontrun{
# phyloseq object
data("Zeybel_2022_gut")
run_betadisper(Zeybel_2022_gut,
   level = "Phylum",
   variable = "LiverFatClass",
   method = "bray")

# SummarizedExperiment object
data("Zeybel_2022_protein")
run_betadisper(Zeybel_2022_protein,
   variable = "LiverFatClass",
   method = "bray")
}

}
\references{
Anderson, Marti J., Kari E. Ellingsen, and Brian H. McArdle.
"Multivariate dispersion as a measure of beta diversity."
Ecology letters 9.6 (2006): 683-693.
}
\author{
Created by Hua Zou (5/15/2022 Shenzhen China)
}
