% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{run_CI}
\alias{run_CI}
\title{Confidence Interval}
\usage{
run_CI(x, ci = 0.95)
}
\arguments{
\item{x}{(Required). vector. a vector of data.}

\item{ci}{(optional). numeric. confidence interval to be calculated.}
}
\value{
\item{upper}{Upper bound of interval.}
\item{mean}{Mean of data.}
\item{lower}{Lower bound of interval.}
}
\description{
Calculates the confidence interval of a vector of data.
}
\examples{
run_CI(rnorm(100))

}
\references{
https://github.com/cran/Rmisc/
}
\author{
Created by Hua Zou (5/19/2022 Shenzhen China)
}
