% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{run_group.STDERR}
\alias{run_group.STDERR}
\title{Group Standard Error Interval}
\usage{
run_group.STDERR(x, data)
}
\arguments{
\item{x}{(Required). an \code{aggregate} compatible formula.}

\item{data}{(Required). a data frame (or list) from which the variables in formula should be taken.}
}
\value{
A data frame consisting of one column for each grouping factor plus
three columns for the upper bound, mean and lower bound of the standard error
interval for each level of the grouping factor.
}
\description{
Calculates the standard error interval of grouped data.
}
\examples{
run_group.STDERR(Sepal.Length~Species, iris)

}
\references{
https://github.com/cran/Rmisc/
}
\author{
Created by Hua Zou (5/19/2022 Shenzhen China)
}
