% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_MANTEL.R
\name{run_MANTEL}
\alias{run_MANTEL}
\title{Mantel Test (MANTEL)}
\usage{
run_MANTEL(
   object,
   level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
   y_variables,
   z_variables = NULL,
   norm = TRUE,
   method = c("mantel", "mantel.partial", "mantel.randtest", "mantel.rtest"),
   method_cor = c("pearson", "spearman", "kendall"),
   method_dist = c("bray", "euclidean", "jaccard"),
   seedNum = 123,
   alpha = 0.5)
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object.}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}

\item{y_variables}{(Required). vector. variables for a control matrix Y.}

\item{z_variables}{(Optional). vector. variables for a control matrix Z (default: NULL).}

\item{norm}{(Optional). logical. whether to norm y and z matrix into unit by
\code{scale} (default: TRUE).}

\item{method}{(Required). character. Mantel test method, "mantel" or
"mantel.partial" or "mantel.randtest", "mantel.rtest" (default: "mantel").}

\item{method_cor}{(Optional). character. Correlation method, as
accepted by cor: "pearson", "spearman" or "kendall" (default: "spearman").}

\item{method_dist}{(Required). character. methods for three(X, Y, Z matrix)
distance matrices. Provide one of the currently supported
options. See \code{distanceMethodList} for a detailed list of the supported options
and links to accompanying documentation. Options include:
\itemize{
\item "unifrac" : unweighted UniFrac distance.
\item "wunifrac": weighted-UniFrac distance.
\item "GUniFrac": The variance-adjusted weighted UniFrac distances (default: alpha=0.5).
\item "bray": bray crutis distance.
\item "dpcoa": sample-wise distance used in Double Principle Coordinate Analysis.
\item "jsd": Jensen-Shannon Divergence.
Alternatively, you can provide a character string that defines a custom
distance method, if it has the form described in \code{designdist}
(default: c("bray", "euclidean", "jaccard")).
}}

\item{seedNum}{(Optional). numeric. specify seeds for reproduction (default: 123).}

\item{alpha}{(Optional). numeric. the parameter for "GUniFrac" controlling
weight on abundant lineages (default: 0.5).}
}
\value{
A Mantel test model.
}
\description{
The \code{run_MANTEL} function is used to identify the association between
the community and environmental variables.
}
\details{
The \code{run_MANTEL} function is used to test the correlation between two
distance matrices, calculating the Z statistic between community distance
and variables distance to determine the significance. While the Mantel test
is used to compare between two distance (dissimilarity) matrices, such as
X and Y, the partial Mantel test is used to estimate the correlation
between these two matrices, while controlling for the effect of
a control matrix Z. It can be applied to both
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} and \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object.
}
\examples{

\dontrun{
data("enterotypes_arumugam")
run_MANTEL(enterotypes_arumugam,
          y_variables = c("Enterotype", "Clinical.Status"),
          z_variables = c("Nationality", "Gender"),
          norm = FALSE,
          method = "mantel.partial",
          method_dist = c("bray", "euclidean", "jaccard"))
}

}
\references{
Mantel, Nathan. "The detection of disease clustering and
a generalized regression approach." Cancer research 27.2 Part 1
(1967): 209-220.
}
\author{
Created by Hua Zou (5/15/2022 Shenzhen China)
}
