% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_metabolome.R
\name{run_metabolomeDA}
\alias{run_metabolomeDA}
\title{Perform differential analysis on metabolomic data}
\usage{
run_metabolomeDA(
    object_raw = NULL,
    object_norm = NULL,
    variable,
    variable_name,
    DA_method = c("all", "fc", "vip", "t"),
    method_VIP = c("PLS", "OPLS"),
    cutoff_prev = 0.1,
    p_adjust = c("none", "fdr", "bonferroni", "holm",
                 "hochberg", "hommel", "BH", "BY"))
}
\arguments{
\item{object_raw}{(Optional). a \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}}
object with raw intensity values (default: NULL).}

\item{object_norm}{(Optional). a \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}}
object with normalized intensity values (default: NULL).}

\item{variable}{(Required). character. grouping variable for test.}

\item{variable_name}{(Required). two characters. variable' names.}

\item{DA_method}{(Optional). character. method for differential analysis
(default: "all"). options include:
\itemize{
\item "all": three methods.
\item "fc": foldchange analysis.
\item "vip": VIP analysis.
\item "t": t-test analysis.
}}

\item{method_VIP}{(Optional). character. method for VIP (default: "PLS"),
options include:
\itemize{
\item "PLS": Partial Least Squares Discriminant Analysis.
\item "OPLS": Orthogonal Partial Least Square Discriminant Analysis.
}}

\item{cutoff_prev}{(Optional). Numeric.
the Prevalence threshold (default: 0.1).}

\item{p_adjust}{(Optional). character. method to adjust p-values by.
Options include "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none". See \code{\link[stats:p.adjust]{stats::p.adjust()}}
for more details (default: "BH").}
}
\value{
A data frame of the differential results.
}
\description{
Differential expression analysis contains foldchange analysis, VIP and t-test
on metabolomic data.
}
\details{
To identify the potential significant metabolites is important
in metabolomics. Here, we use three methods to obtain the results:
\enumerate{
\item FoldChange: the raw intensity values;
\item VIP (Variable important in projection) by PLS-DA:
the normalized intensity values;
\item T-test: the normalized intensity values.
}

Combining all the results and providing more choice for users to determine
the differential metabolites (Recommend: Log2FoldChange and AdjustedPvalue).
}
\examples{

data("Zeybel_2022_protein")
Zeybel_2022_protein_imput <- impute_abundance(
      Zeybel_2022_protein,
      group = "LiverFatClass",
      method = "knn")
Zeybel_2022_protein_norm <- scale_variables(
      Zeybel_2022_protein_imput,
      method == "zscore")
DA_results <- run_metabolomeDA(
  object_raw = Zeybel_2022_protein,
  object_norm = Zeybel_2022_protein_norm,
  variable = "LiverFatClass",
  variable_name = c("None", "Severe"))

\dontrun{

}

}
\author{
Created by Hua Zou (12/2/2023 Shenzhen China)
}
