% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_MRPP.R
\name{run_MRPP}
\alias{run_MRPP}
\title{Multi-response Permutation Procedures (MRPP)}
\usage{
run_MRPP(
   object,
   level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
   variable,
   method = c("unifrac", "wunifrac", "GUniFrac", "bray", "dpcoa", "jsd"),
   seedNum = 123,
   alpha = 0.5)
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object.}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}

\item{variable}{(Required). character. grouping variable for test.}

\item{method}{(Optional). character. Provide one of the currently supported
options. See \code{distanceMethodList} for a detailed list of the supported options
and links to accompanying documentation. Options include:
\itemize{
\item "unifrac" : unweighted UniFrac distance.
\item "wunifrac": weighted-UniFrac distance.
\item "GUniFrac": The variance-adjusted weighted UniFrac distances (default: alpha=0.5).
\item "bray": bray crutis distance.
\item "dpcoa": sample-wise distance used in Double Principle Coordinate Analysis.
\item "jsd": Jensen-Shannon Divergence.
Alternatively, you can provide a character string that defines a custom
distance method, if it has the form described in \code{designdist} (default: "bray").
}}

\item{seedNum}{(Optional). numeric. specify seeds for reproduction (default: 123).}

\item{alpha}{(Optional). numeric. the parameter for "GUniFrac" controlling
weight on abundant lineages (default: 0.5).}
}
\value{
A MRPP model.
}
\description{
The \code{run_MRPP} function is used to identify the association between
the community and environmental variables by average distance.
}
\details{
The \code{run_MRPP} function is used to identify the association between
the community and environmental variables by average distance, applying the
distance in profile and calculating the delta statistic between community and
variable by permutation test to determine the significance. It can be applied
to both \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} and \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object.
}
\examples{

\dontrun{
data("enterotypes_arumugam")
run_MRPP(enterotypes_arumugam, variable = "Enterotype", method = "bray")
}

}
\references{
Mielke Jr, Paul W. "The application of multivariate permutation
methods based on distance functions in the earth sciences."
Earth-Science Reviews 31.1 (1991): 55-71.
}
\author{
Created by Hua Zou (5/15/2022 Shenzhen China)
}
