% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{run_OddRatio}
\alias{run_OddRatio}
\title{95\% confidential interval Odds Ratio}
\usage{
run_OddRatio(datx, daty, GroupName)
}
\arguments{
\item{datx}{(Required). Data.frame. Metadata with SampleID and
Group information.}

\item{daty}{(Required). Data.frame. Profile with SampleID,
ordered by datx (Row->Features; Column->SampleID).}

\item{GroupName}{(Required). Character. the contrast group.}
}
\value{
95\% confidential interval Odds Ratio per feature
}
\description{
Odds ratio shows the ratio between the probability of Group1 and Group2,
indicating the risk between groups. \code{run_OddRatio} is used glm with
scaled values to calculate coefficients of groups to obtain the
95\% confidential interval.
}
\author{
Created by Hua Zou (12/13/2021 Guangzhou China)
}
