% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_ordination.R
\name{run_ord}
\alias{run_ord}
\title{Ordination for microbiota data}
\usage{
run_ord(
    object,
    level = NULL,
    variable,
    transform = c("identity", "log10", "log10p",
                  "SquareRoot", "CubicRoot", "logit"),
    norm = c("none", "rarefy", "TSS", "TMM",
             "RLE", "CSS", "CLR", "CPM"),
    method = c("PCA", "PCoA", "tSNE", "UMAP", "NMDS",
               "CA", "RDA", "CCA", "CAP"),
    distance = c("bray", "unifrac", "wunifrac",
                 "GUniFrac", "dpcoa", "jsd"),
    para = list(Perplexity = NULL,
                Y_vars = NULL,
                Z_vars = NULL,
                scale = TRUE,
                center = TRUE,),
    ...)
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object.}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}

\item{variable}{(Required). character. grouping variable for test.}

\item{transform}{character, the methods used to transform the microbial
abundance. See \code{\link[=transform_abundances]{transform_abundances()}} for more details. The
options include:
\itemize{
\item "identity", return the original data without any transformation
(default).
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
\item "SquareRoot", the transformation is \verb{Square Root}.
\item "CubicRoot", the transformation is \verb{Cubic Root}.
\item "logit", the transformation is \verb{Zero-inflated Logit Transformation}
(Does not work well for microbiome data).
}}

\item{norm}{the methods used to normalize the microbial abundance data. See
\code{\link[=normalize]{normalize()}} for more details.
Options include:
\itemize{
\item "none": do not normalize.
\item "rarefy": random subsampling counts to the smallest library size in the
data set.
\item "TMM": trimmed mean of m-values. First, a sample is chosen as reference.
The scaling factor is then derived using a weighted trimmed mean over the
differences of the log-transformed gene-count fold-change between the
sample and the reference.
\item "RLE", relative log expression, RLE uses a pseudo-reference calculated
using the geometric mean of the gene-specific abundances over all
samples. The scaling factors are then calculated as the median of the
gene counts ratios between the samples and the reference.
\item "CSS": cumulative sum scaling, calculates scaling factors as the
cumulative sum of gene abundances up to a data-derived threshold.
}}

\item{method}{(Optional). character. Ordination method (default: "PCoA"),
options include:
\itemize{
\item "PCA": Principal Component Analysis.
\item "PCoA": Principal Coordinate Analysis.
\item "tSNE": t-distributed stochastic neighbor embedding.
\item "UMAP": Uniform Manifold Approximation and Projection.
\item "NMDS": Non-metric Multidimensional Scaling.
}}

\item{distance}{(Optional). character. Provide one of the currently supported
options. See \code{vegan::vegdist} for a detailed list of the supported options
and links to accompanying documentation (default: "bray"). Options include:
\itemize{
\item "bray": bray crutis distance.
\item "unifrac" : unweighted UniFrac distance.
\item "wunifrac": weighted-UniFrac distance.
\item "GUniFrac": The variance-adjusted weighted UniFrac distances (default: alpha=0.5).
\item "dpcoa": sample-wise distance used in Double Principle Coordinate Analysis.
\item "jsd": Jensen-Shannon Divergence.
Alternatively, you can provide a character string that defines a custom
distance method, if it has the form described in \code{designdist}.
}}

\item{para}{(Optional). list. the additional parameters for methods.
\itemize{
\item "Perplexity": numeric; Perplexity parameter (should not be bigger than
3 perplexity < nrow(X) - 1.
\item "Y_vars": Constraining matrix, typically of environmental variables.
\item "Z_vars": Conditioning matrix, the effect of which is removed ("partial out")
before next step.
\item "scale": Scale features to unit variance (like correlations).
\item "center": Scale features to unit variance (like correlations).
}}

\item{...}{(Optional). additional parameters.}
}
\value{
A list of the ordination's results.
}
\description{
The primary goal of ordination was considered “exploratory” (Gauch 1982a, b),
with the introduction of canonical correspondence analysis (CCA), ordination
has gone beyond mere “exploratory” analysis (ter Braak 1985) and become
hypothesis testing as well.
}
\details{
The primary aim of ordination is to represent multiple samples (subjects) in
a reduced number of orthogonal (i.e., independent) axes,
where the total number of axes is less than or equal to the number of samples
}
\examples{

\dontrun{

# phyloseq object
data("Zeybel_2022_gut")
ps_zeybel <- summarize_taxa(Zeybel_2022_gut, level = "Genus")
ord_result <- run_ord(
  object = ps_zeybel,
  variable = "LiverFatClass",
  method = "PCoA")

# SummarizedExperiment object
data("Zeybel_2022_protein")
Zeybel_2022_protein_imp <- impute_abundance(
  Zeybel_2022_protein,
  group = "LiverFatClass",
  ZerosAsNA = TRUE,
  RemoveNA = TRUE,
  cutoff = 20,
  method = "knn")
ord_result <- run_ord(
  object = Zeybel_2022_protein_imp,
  variable = "LiverFatClass",
  method = "PCA")

}

}
\references{
Xia, Y., Sun, J., & Chen, D. G. (2018). Statistical analysis of microbiome
data with R (Vol. 847). Singapore: Springer.
}
\author{
Created by Hua Zou (8/9/2023 Shenzhen China)
}
