% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_PERMANOVA.R
\name{run_PERMANOVA}
\alias{run_PERMANOVA}
\title{Permutational multivariate analysis of variance (PERMANOVA)}
\usage{
run_PERMANOVA(object,
                     variables,
                     method,
                     seedNum,
                     alpha)
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object.}

\item{variables}{(Optional). vector. variables for test.}

\item{method}{(Optional). character. Provide one of the currently supported
options. See \code{distanceMethodList} for a detailed list of the supported options
and links to accompanying documentation. Options include:
\itemize{
\item "unifrac" : unweighted UniFrac distance.
\item "wunifrac": weighted-UniFrac distance.
\item "GUniFrac": The variance-adjusted weighted UniFrac distances (default: alpha=0.5).
\item "bray": bray crutis distance.
\item "dpcoa": sample-wise distance used in Double Principle Coordinate Analysis.
\item "jsd": Jensen-Shannon Divergence.
Alternatively, you can provide a character string that defines a custom
distance method, if it has the form described in \code{designdist} (default: "bray").
}}

\item{seedNum}{(Optional). numeric. specify seeds for reproduction (default: 123).}

\item{alpha}{(Optional). numeric. the parameter for "GUniFrac" controlling
weight on abundant lineages (default: 0.5).}
}
\value{
A data.frame of PERMANOVA result.
}
\description{
The \code{run_PERMANOVA} function is used to identify the association between
the community and environmental variables.
}
\details{
The \code{run_PERMANOVA} function is used to identify the association between
the community and environmental variables, applying the distance in profile
and calculating the F statistic between community and variable by permutation
test to determine the significance. It can be applied to both
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} and \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object.
}
\examples{

\dontrun{
data("enterotypes_arumugam")
run_PERMANOVA(enterotypes_arumugam, method = "bray")
}

}
\references{
Anderson, Marti J. "Permutational multivariate analysis of
variance." Department of Statistics, University of Auckland,
Auckland 26 (2005): 32-46.
}
\author{
Created by Hua Zou (5/14/2022 Shenzhen China)
}
