% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_rarecurve.R
\name{run_rarecurve}
\alias{run_rarecurve}
\title{Calculating result of rarefaction curve}
\usage{
run_rarecurve(
   object,
   level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
   index_name = c("Observed", "Chao1", "ACE", "Shannon", "Simpson",
                 "InvSimpson", "Fisher", "Evenness", "readsNums"),
   chunks = 200,
   rng_seed = 123)
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object.}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}

\item{index_name}{(Required). character.
alpha-diversity indices (default: "Observed").
Values must be among those supported:
c("Observed", "Chao1", "ACE", "Shannon", "Simpson",
"InvSimpson", "Fisher", "Evenness", "readsNums").
"Chao1", "ACE", "Fisher" and "readsNums" only supported by
counts matrix (integers).}

\item{chunks}{(Optional). integer. the number of index in a sample (default: 200).}

\item{rng_seed}{(Optional). integer. random seed (default: 123).}
}
\value{
an object, which could be applied for rarefaction curve plotting.
}
\description{
The function is to result of rarefaction curve
}
\examples{

\dontrun{
# phyloseq object
data("enterotypes_arumugam")
run_rarecurve(
   object = enterotypes_arumugam,
   level = "Phylum",
   index_name = "Observed")
}

}
\author{
Created by Hua Zou (5/14/2022 Shenzhen China)
}
