% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{run_STDERR}
\alias{run_STDERR}
\title{Standard Error}
\usage{
run_STDERR(x)
}
\arguments{
\item{x}{(Required). vector. a vector of data.}
}
\value{
\item{upper}{Upper bound of interval.}
\item{mean}{Mean of data.}
\item{lower}{Lower bound of interval.}
}
\description{
Calculates the standard error interval of a vector of data
}
\examples{
run_STDERR(rnorm(100))

}
\references{
https://github.com/cran/Rmisc/
}
\author{
Created by Hua Zou (5/19/2022 Shenzhen China)
}
