% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{run_summarySE}
\alias{run_summarySE}
\title{Summarizes data for mean median sd etc}
\usage{
run_summarySE(
  data,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{(Required). a data frame.}

\item{measurevar}{(Required). character. the name of a column that contains the variable to be summariezed.}

\item{groupvars}{(Required). vector. a vector containing names of columns that contain grouping variables.}

\item{na.rm}{(Required). logical. a boolean that indicates whether to ignore NA's (default: FALSE).}

\item{conf.interval}{(Required). numeric. the percent range of the confidence interval (default: 0.95).}

\item{.drop}{(Required). logical. should combinations of variables that do not appear in the input data be
preserved (FALSE) or dropped (default: TRUE).}
}
\value{
a data frame with count, mean, standard deviation,
standard error of the mean, and confidence interval (default 95\%).
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95\%).
}
\examples{

\dontrun{
library(dose)
run_summarySE(ToothGrowth, measurevar="len", groupvars=c("supp", "dose"))
}

}
\references{
https://github.com/cran/Rmisc/
}
\author{
Created by Hua Zou (5/19/2022 Shenzhen China)
}
