% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_normalization.R
\name{scale_variables}
\alias{scale_variables}
\title{Data scaling adjusts each variable/feature}
\usage{
scale_variables(
   object,
   level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
   method = c("none", "mean_center", "zscore",
            "pareto", "range")
   )
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object.}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}

\item{method}{(Optional). character. scaling methods. Options are:
\itemize{
\item "none", return the original data
\item "mean_center": values minus mean statistic.
\item "zscore": mean-centered and divided by the
standard deviation of each variable.
\item "pareto": mean-centered and divided by the square root of
the standard deviation of each variable.
\item "range": mean-centered and divided by the range of each variable.
(default: "none").
}}
}
\value{
A \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object with cleaned data.
}
\description{
Data Normalization not only normalizes data by samples, but also
scales data by variable/feature. Data scaling adjusts each variable/feature
by a scaling factor computed based on the dispersion of the variable. The
former is to change data distribution per sample and the latter is to put
variable/feature into same distribution.
}
\examples{

\dontrun{
# phyloseq object
data("enterotypes_arumugam")
scale_variables(
  object = enterotypes_arumugam,
  level = "Phylum",
  method = "pareto")

# SummarizedExperiment object
data("Zeybel_2022_protein")
scale_variables(
  Zeybel_2022_protein,
  method = "zscore")
}

}
\author{
Created by Hua Zou (12/02/2022 Shenzhen China)
}
