% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trim_prevalence.R
\name{trim_prevalence}
\alias{trim_prevalence}
\title{Trimming samples or features whose prevalence is less than threshold}
\usage{
trim_prevalence(
    object,
    level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
    cutoff = 0.1,
    trim = c("none", "both", "feature", "sample"))
}
\arguments{
\item{object}{(Required). a \code{\link{matrix}}, \code{\linkS4class{otu_table}},
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or \code{\linkS4class{SummarizedExperiment}}}

\item{level}{(Optional). character. taxonomic level to summarize,
default the top level rank of the \code{ps}. taxonomic level(Kingdom, Phylum,
Class, Order, Family, Genus, Species, Strains; default: NULL).}

\item{cutoff}{(Optional). Numeric. the Prevalence threshold (default: 0.1).}

\item{trim}{(Optional). Character. trimming to apply, the options include:
\itemize{
\item "none", return the original data without any actions.
\item "both", prevalence of features and samples more than cutoff.
\item "feature", prevalence of features more than cutoff.
\item "sample", prevalence of samples more than cutoff.
}}
}
\value{
A trimed \code{object} whose prevalence of features or samples more than cutoff.
}
\description{
trim samples or features in \code{profile} by Prevalence,
which means the samples or features will be discarded if they could not pass the cutoff.
}
\examples{
\donttest{
# phyloseq object
data("Zeybel_2022_gut")
trim_prevalence(
  Zeybel_2022_gut,
  level = "Phylum",
  cutoff = 0.1,
  trim = "feature")

# SummarizedExperiment object
data("Zeybel_2022_protein")
trim_prevalence(
  Zeybel_2022_protein,
  level = NULL,
  cutoff = 0.1,
  trim = "feature")
}

}
\author{
Created by Hua Zou (11/30/2021 Shenzhen China)
}
