% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_filter_abundance.R
\name{filter_abundance}
\alias{filter_abundance}
\title{Filtering taxa who is low relative abundance or unclassified}
\usage{
filter_abundance(
    object,
    level = NULL,
    cutoff_mean = c(100, 1e-04),
    cutoff_one = c(1000, 1e-03),
    unclass = TRUE)
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq]{phyloseq-class}} object.}

\item{level}{(Optional). character. taxonomic level to summarize,
default the top level rank of the \code{ps}. taxonomic level(Kingdom, Phylum,
Class, Order, Family, Genus, Species, Strains; default: NULL).}

\item{cutoff_mean}{(Optional). numeric. Threshold for Mean
relative abundance all samples (default, \code{1e-04}).}

\item{cutoff_one}{(Optional). numeric. Threshold for Minimum
relative abundance at least one sample (default, \code{1e-03}).}

\item{unclass}{(Optional). logical. whether to filter the unclassified taxa (default \code{TRUE}).}
}
\value{
a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object, where each row represents a
taxa, and each col represents the taxa abundance of each sample.
}
\description{
whether to filter the low relative abundance or unclassified
taxa by the threshold. Here, we choose the following criterion:
\enumerate{
\item Taxa more than Mean relative abundance across all samples: 0.0001 (1e-04);
\item Taxa more than Minimum relative abundance at least one sample: 0.001 (1e-03).
}
}
\examples{

\dontrun{
 data("enterotypes_arumugam")

 # absolute abundance
 ps <- filter_abundance(
   object = enterotypes_arumugam,
   level = NULL,
   cutoff_mean = 100,
   cutoff_one = 1000,
   unclass = FALSE)

 # relative abundance
 enterotypes_arumugam_rb <- normalize(enterotypes_arumugam, method = "TSS")
 ps <- filter_abundance(
   object = enterotypes_arumugam_rb,
   level = "Phylum",
   cutoff_mean = 1e-04,
   cutoff_one = 1e-03,
   unclass = TRUE)

}

}
\references{
Thingholm, Louise B., et al. "Obese individuals with and without
type 2 diabetes show different gut microbial functional capacity and
composition." Cell host & microbe 26.2 (2019): 252-264.
}
