% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_imputation.R
\name{impute_abundance}
\alias{impute_abundance}
\title{Imputation methods on missing value}
\usage{
impute_abundance(
   object,
   level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
   group_name,
   ZerosAsNA = FALSE,
   RemoveNA = TRUE,
   cutoff = 20,
   method = c("none", "LOD", "half_min", "median",
       "mean", "min", "knn", "rf",
       "global_mean", "svd", "QRILC"),
   LOD = NULL
   )
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object.}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}

\item{group_name}{(Required). character. group for determining missing values.}

\item{ZerosAsNA}{(Optional). logical. zeros in the data are missing
values (default: FALSE).}

\item{RemoveNA}{(Optional). logical. those features with more than selected
cutoff missing values in each group have to be removed (default: TRUE).}

\item{cutoff}{(Optional). numeric. percentage of missing values allowed in
each group. If one of the groups have less missing values than selected
cutoff value, these feature will not be removed.}

\item{method}{(Optional). character. Imputation method. Options are:
\itemize{
\item "none": all missing values will be replaced by zero.
\item "LOD": specific Limit Of Detection which provides by user.
\item "half_min": half minimal values across samples except zero.
\item "median": median values across samples except zero.
\item "mean": mean values across samples except zero.
\item "min": minimal values across samples except zero.
\item "knn": k-nearest neighbors samples.
\item "rf": nonparametric missing value imputation using Random Forest.
\item "global_mean": a normal distribution with a mean that is down-shifted from
the sample mean and a standard deviation that is a
fraction of the standard deviation of the sample distribution.
\item "svd": missing values imputation based Singular value decomposition.
\item "QRILC": missing values imputation based quantile regression.
(default: "none").
}}

\item{LOD}{(Optional). Numeric. limit of detection (default: NULL).}
}
\value{
A \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object with cleaned data.
}
\description{
This function offers different methods to impute missing values in data.
}
\examples{

\dontrun{
# phyloseq object
data("Zeybel_2022_gut")
impute_abundance(
  Zeybel_2022_gut,
  level = "Phylum",
  group_name = "LiverFatClass",
  ZerosAsNA = TRUE,
  RemoveNA = TRUE,
  cutoff = 20,
  method = "knn")

# SummarizedExperiment object
data("Zeybel_2022_protein")
impute_abundance(
  Zeybel_2022_protein,
  level = "Phylum",
  group_name = "LiverFatClass",
  ZerosAsNA = TRUE,
  RemoveNA = TRUE,
  cutoff = 20,
  method = "knn")
}

}
\references{
Armitage, E. G., Godzien, J., Alonso‐Herranz, V., López‐Gonzálvez,
Á., & Barbas, C. (2015). Missing value imputation strategies for metabolomics
data. Electrophoresis, 36(24), 3050-3060.
}
\author{
Created by Pol Castellano-Escuder;
Modified by Hua Zou (12/02/2022 Shenzhen China)
}
