% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_PERMANOVA.R
\name{run_PERMANOVA}
\alias{run_PERMANOVA}
\title{Permutational multivariate analysis of variance (PERMANOVA)}
\usage{
run_PERMANOVA(
   object,
   level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"),
   variables = "all",
   method = c("unifrac", "wunifrac", "GUniFrac", "bray", "dpcoa", "jsd"),
   mode = c("one", "all"),
   seedNum = 123,
   alpha = 0.5)
}
\arguments{
\item{object}{(Required). a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} or
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object.}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}

\item{variables}{(Optional). vector. variables for test (default: all).}

\item{method}{(Optional). character. Provide one of the currently supported
options. See \code{distanceMethodList} for a detailed list of the supported options
and links to accompanying documentation. Options include:
\itemize{
\item "bray": bray crutis distance.
\item "unifrac": unweighted UniFrac distance.
\item "wunifrac": weighted-UniFrac distance.
\item "GUniFrac": The variance-adjusted weighted UniFrac distances (default: alpha=0.5).
\item "dpcoa": sample-wise distance used in Double Principle Coordinate Analysis.
\item "jsd": Jensen-Shannon Divergence.
Alternatively, you can provide a character string that defines a custom
distance method, if it has the form described in \code{designdist} (default: "bray").
}}

\item{mode}{(Optional). character. Since there are missing values
in some columns, providing two test model:
\itemize{
\item "one": test on one variable per time.
\item "all": test on all variables once.
(default: "one").
}}

\item{seedNum}{(Optional). numeric. specify seeds for reproduction (default: 123).}

\item{alpha}{(Optional). numeric. the parameter for "GUniFrac" controlling
weight on abundant lineages (default: 0.5).}
}
\value{
A data.frame of PERMANOVA result.
}
\description{
The \code{run_PERMANOVA} function is used to identify the association between
the community and environmental variables.
}
\details{
The \code{run_PERMANOVA} function is used to identify the association between
the community and environmental variables, applying the distance in profile
and calculating the F statistic between community and variable by permutation
test to determine the significance. It can be applied to both
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} and \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object.
}
\examples{

\dontrun{
# phyloseq object
data("Zeybel_2022_gut")
run_PERMANOVA(Zeybel_2022_gut,
          variables = c("LiverFatClass", "Liver_fat", "Gender"),
          mode = "one",
          method = "bray")

# SummarizedExperiment object
data("Zeybel_2022_protein")
se_impute <- impute_abundance(
  object = Zeybel_2022_protein,
  group = "LiverFatClass",
  ZerosAsNA = TRUE,
  RemoveNA = TRUE,
  cutoff = 20,
  method = "knn")
run_PERMANOVA(se_impute,
          variables = c("LiverFatClass", "Liver_fat"),
          mode = "all",
          method = "bray")
}

}
\references{
Anderson, Marti J. "Permutational multivariate analysis of
variance."Department of Statistics, University of Auckland,
Auckland 26 (2005): 32-46.
}
\author{
Created by Hua Zou (5/14/2022 Shenzhen China)
}
