% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_transform.R
\name{transform_abundances}
\alias{transform_abundances}
\title{Transform the abundances in \code{profile} sample by sample}
\usage{
transform_abundances(
   object,
   transform = c("identity", "log10", "log10p",
            "log2", "log2p",
            "SquareRoot", "CubicRoot", "logit"),
   level = c(NULL, "Kingdom", "Phylum", "Class",
           "Order", "Family", "Genus",
           "Species", "Strain", "unique"))
}
\arguments{
\item{object}{(Required). a \code{\linkS4class{otu_table}}, \code{\linkS4class{phyloseq}},
\code{\linkS4class{SummarizedExperiment}} or \code{\linkS4class{microbiomeMarker}}.}

\item{transform}{transformation to apply, the options inclulde:
\itemize{
\item "identity", return the original data without any transformation.
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
\item "log2", the transformation is \code{log2(object)}, and if the data contains
zeros the transformation is \code{log2(1 + object)}.
\item "log2p", the transformation is \code{log2(1 + object)}.
\item "SquareRoot", the transformation is \verb{Square Root}.
\item "CubicRoot", the transformation is \verb{Cubic Root}.
\item "logit", the transformation is \verb{Zero-inflated Logit Transformation}
(Does not work well for microbiome data).
}}

\item{level}{(Optional). character. Summarization
level (from \code{rank_names(pseq)}, default: NULL).}
}
\value{
A object matches the class of argument \code{object} with the transformed
\code{otu_table}.
}
\description{
Transform the abundances in \code{profile} sample by sample, which means
the values of each sample will be transformed individually. It will
change the data distribution of features. For instance, transforming
left skew distribution into nearly norml distribution by using log function.
}
\examples{
\dontrun{
# phyloseq object
data("Zeybel_2022_gut")
transform_abundances(
  Zeybel_2022_gut,
  level = "Phylum",
  transform = "log10p")

# SummarizedExperiment object
data("Zeybel_2022_protein")
transform_abundances(
  Zeybel_2022_protein,
  transform = "log10p")
}

}
\seealso{
\code{\link[=abundances]{abundances()}}
}
\author{
Created by Yang Cao; modified by Hua Zou (11/30/2022 Shenzhen China)
}
