% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trim_prevalence.R
\name{trim_prevalence}
\alias{trim_prevalence}
\title{Trimming samples or taxa whose prevalence is less than threshold}
\usage{
trim_prevalence(
    object,
    level = NULL,
    cutoff = 0.1,
    trim = c("none", "both", "feature", "sample"))
}
\arguments{
\item{object}{(Required). a \code{\link{matrix}} or \code{\linkS4class{otu_table}} or \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}.}

\item{level}{(Optional). character. taxonomic level to summarize,
default the top level rank of the \code{ps}. taxonomic level(Kingdom, Phylum,
Class, Order, Family, Genus, Species, Strains; default: NULL).}

\item{cutoff}{(Optional). Numeric. the Prevalence threshold (default: 0.1).}

\item{trim}{(Optional). Character. trimming to apply, the options include:
\itemize{
\item "none", return the original data without any actions.
\item "both", prevalence of taxa and samples more than cutoff.
\item "feature", prevalence of taxa more than cutoff.
\item "sample", prevalence of samples more than cutoff.
}}
}
\value{
A trimed \code{object} whose prevalence of taxa or samples more than cutoff.
}
\description{
trim samples or taxa in \code{otu_table} by Prevalence,
which means the samples or taxa will be discarded if they could not pass the cutoff.
}
\examples{
\donttest{
 data("enterotypes_arumugam")
 trim_prevalence(object = enterotypes_arumugam,
   cutoff = 0.1, trim = "feature")
}

}
\author{
Created by Hua Zou (11/30/2021 Shenzhen China)
}
